webapp/requirements.txt
flask
requests
black

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
efficientnet_pytorch
tensorboard>=2.4.1

seaborn>=0.11.0
pandas

thop  # FLOPs computation
